This is for maximum pariwise product.

+++ Maximum_Pairwise_Product.c
@@ -9,7 +9,7 @@ long long MaxPairwiseProductFast(int numbers[], int n) {
   for (int j = 0; j < n; ++j)
     if ((j!= max_index1) && ((max_index2 == -1) || (numbers[j] > numbers[max_index2])))
       max_index2 = j;
-  long long result=numbers[max_index1] * numbers[max_index2];
+  long long result=((long long)(numbers[max_index1])) * numbers[max_index2];
   return (result);
 }
 
@@ -22,6 +22,6 @@ int main() {
   }
 
   long long result = MaxPairwiseProductFast(numbers, n);
-  printf("%lld\n", result);
+  printf("%lli\n", result);
   return 0;
 }
